chore(tests): move web service tests to phpunit
